Working alarm from Arduino source
Added the ability for the web page to parse the alarms from the arduino.
If one doesn't quite match the ones in the hidden field, it will add it
to the hidden field

This means that a separate browser will be updated with timers set by
the source browser.
The timer title will not be populated.